refactor: move readVector
