+++ src/viperleed/calc/lib/string_utils.py
@@ -17,6 +17,8 @@ __license__ = 'GPLv3+'
 import itertools
 import re
 
+import numpy as np
+
 from viperleed.calc.lib.itertools_utils import batched
 from viperleed.calc.lib.itertools_utils import consecutive_groups
 
@@ -152,6 +154,29 @@ def read_int_range(ranges):
     if not integers:
         raise ValueError(f'No integers in {ranges!r}')
     return integers
+
+
+def readVector(s, ucell=None, defRelaltive=False):
+    """Takes a string 'xyz[f1 f2 f3]', 'abc[f1 f2 f3]' or just '[f1 f2 f3]'
+    and returns the corresponding vector in cartesian coordinates, or None if
+    the string cannot be parsed."""
+    m = re.match(r'\s*(xyz|abc)?\[\s*(?P<v1>[-0-9.]+)\s+'
+                 r'(?P<v2>[-0-9.]+)\s+(?P<v3>[-0-9.]+)\s*\]', s)
+    if not m:
+        return None
+    try:
+        v1 = float(m.group("v1"))
+        v2 = float(m.group("v2"))
+        v3 = float(m.group("v3"))
+    except (ValueError, IndexError):
+        return None
+    if "abc" in s:
+        if ucell is None:
+            return None
+        uct = ucell.T
+        return np.dot((v1, v2, v3), ucell.T)
+    # xyz
+    return np.array([v1, v2, v3])
 
 
 def rsplit_once(string, sep):
